+++ home/linux/gui/base/xdg/mime.nix
@@ -15,8 +15,6 @@
       let
         browser = [
           "zen.desktop"
-          "firefox.desktop"
-          "google-chrome.desktop"
         ];
         editor = [
           "nvim.desktop"